adding alerts with signalr

+++ react-crash-2021.app/Data/Repositories/IAlertRepository.cs
@@ -10,7 +10,7 @@ namespace react_crash_2021.Data.Repositories
     {
         public Task<IEnumerable<alert>> GetAlertsByUser(Guid id);
         public Task<int> DeleteAlert(long id);
-        public Task<int> CreateAlert(alert a);
+        public Task<long> CreateAlert(alert a);
         public Task<int> CreateAlerts(IEnumerable<alert> a);
     }
 }
